optimize_crossword/main.py
import yaml

def load_config(config_file: str):
    with open(config_file, 'r') as f:
        config = yaml.safe_load(f)
    return config
# Import necessary modules and implement your application code
def main():
    # Your main application logic
    print("Welcome to Optimize Crossword!")

if __name__ == "__main__":
    main()
